requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# WebSocket and real-time
websockets>=11.0.0
streamlit-autorefresh>=0.3.0

# Machine Learning
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data generation and utilities
faker>=19.0.0
requests>=2.31.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3 